+++ docs/compare.dart
@@ -34,10 +34,16 @@ void showExamples() {
     var tr = document.createElement('tr') as TableRowElement;
     // TeX input
     var cell1 = document.createElement('td') as TableCellElement;
-    cell1.innerHtml = '<pre><code>$example</code></pre>';
     cell1.style.maxWidth = "250px";
-    cell1.style.wordBreak = "break-all";
+    //cell1.style.wordBreak = "break-all";
     tr.append(cell1);
+
+    var s = example;
+    var div = document.createElement('div') as DivElement;
+    div.innerHtml = '<pre><code>$s</code></pre>';
+    div.style.overflowX = "scroll";
+    cell1.append(div);
+
     for (var i = 0; i < 2; i++) {
       // Dart TeX
       var cell2 = document.createElement('td') as TableCellElement;